Make a very simple plot to visually represent how full and empty freezer is at a given time

Make a very simple plot to visually represent how full and empty freezer is at a given time

Usage

freezer_fullness_graph(used = NULL)

Arguments

used

Percent of freezer currently occupied

Value

A ggplot2 plot