Merge PTF kit projection counts with number of tubes in each kit

Merge PTF kit projection counts with number of tubes in each kit

Usage

merge_ptf_tube_counts(x = NULL, y = NULL)

Arguments

x

PTF projected kit counts

y

Number of tubes in each kit

Value

tibble merged of both above

Examples

merge_ptf_tube_counts()
#> # A tibble: 238 × 8
#>    specimen_clean        Specimen   year  pricing_option kit_count specimen_type
#>    <chr>                 <chr>      <chr> <chr>              <int> <chr>        
#>  1 child_cord_blood_core Cord Blood 2029  p2                  2028 Cord Blood   
#>  2 child_cord_blood_core Cord Blood 2029  p2                  2028 Cord Blood   
#>  3 child_cord_blood_core Cord Blood 2024  p1                  2565 Cord Blood   
#>  4 child_cord_blood_core Cord Blood 2024  p1                  2565 Cord Blood   
#>  5 child_cord_blood_core Cord Blood 2030  p1                  3486 Cord Blood   
#>  6 child_cord_blood_core Cord Blood 2030  p1                  3486 Cord Blood   
#>  7 child_cord_blood_core Cord Blood 2025  p1                  4995 Cord Blood   
#>  8 child_cord_blood_core Cord Blood 2025  p1                  4995 Cord Blood   
#>  9 child_cord_blood_core Cord Blood 2024  p2                  2642 Cord Blood   
#> 10 child_cord_blood_core Cord Blood 2024  p2                  2642 Cord Blood   
#> # ℹ 228 more rows
#> # ℹ 2 more variables: tube_size <chr>, tubes_per_kit <dbl>