Load in predicted participant numbers from Measurement Core May 2025

Load in predicted participant numbers from Measurement Core May 2025

Usage

participants_proj(x = "MC_projections_May2025.csv")

Arguments

x

file name

Value

a tibble with this information

Examples

participants_proj()
#> # A tibble: 12 × 12
#>    pt_group  maternal_specimen child_specimen partner_specimen visit y2024 y2025
#>    <chr>                 <int>          <int>            <int> <chr> <int> <int>
#>  1 0-5M                      1              1                0 0_5_…  2442  4306
#>  2 24-35M                    1              1                0 24_3…  1290  1132
#>  3 6-11M                     1              1                0 6_11…   460  4306
#>  4 Delivery                  1              1                0 peri…  2442  4306
#>  5 P-Early                   1              0                0 preg…  6200  6200
#>  6 P-Late                    1              0                0 preg…  4306  6200
#>  7 11-17Y                    0              1                0 11_1…  5155  6183
#>  8 12-23M                    0              1                0 12_2…  1638  2902
#>  9 18-20Y                    0              1                0 18_2…   884   662
#> 10 3-5Y                      0              1                0 3_5_…  3274  3328
#> 11 6-10Y                     0              1                0 6_10…  5489  5480
#> 12 6-11M (P…                 0              0                1 6_11…     1  2368
#> # ℹ 5 more variables: y2026 <int>, y2027 <int>, y2028 <int>, y2029 <int>,
#> #   y2030 <int>