Read in file from Suchi with historical data on ECHO submissions to the BioStore

Read in file from Suchi with historical data on ECHO submissions to the BioStore

Usage

readHistorical(x = "suchi_bam_submissions.csv")

Arguments

x

File name

Value

A tibble with historical data on ECHO submissions to the BioStore

Examples

readHistorical()
#> # A tibble: 45 × 13
#>    Suchi_date date       Orders Containers total_submitted tubes_1.9_ml
#>         <int> <date>      <int>      <int>           <int>        <int>
#>  1   20241008 2024-10-08      1          7             432          240
#>  2   20241015 2024-10-15      1          2             192            0
#>  3   20241031 2024-10-31      2         30            1728         1152
#>  4   20241104 2024-11-04      3         30            2016          864
#>  5   20241105 2024-11-05      4         60            3695         2063
#>  6   20241113 2024-11-13      4         60            3503         2255
#>  7   20241114 2024-11-14      2         28            1260         1179
#>  8   20241118 2024-11-18      9        135            8544         4416
#>  9   20241120 2024-11-20      5         72            4560         2352
#> 10   20241121 2024-11-21      8        112            7293         3456
#> # ℹ 35 more rows
#> # ℹ 7 more variables: tubes_1.0_ml <int>, cumulative_1.0 <int>,
#> #   cumulative_1.9 <int>, cumulative_total <int>, proportion_1.0 <dbl>,
#> #   proportion_1.9 <dbl>, proportion_total <dbl>