Create a zoo ts object from historical data https://cran.r-project.org/web/packages/zoo/index.html

Create a zoo ts object from historical data https://cran.r-project.org/web/packages/zoo/index.html

Usage

zoo_ts(x = readHistorical(), tube = "cumulative_1.0")

Arguments

x

Historical data

tube

tube size, and whether cumulative or not. Options are "tubes_1.0_ml", "tubes_1.9_ml", "cumulative_1.0" or "cumulative_1.9"

Value

zoo ts object

Examples

zoo_ts()
#> 2024-10-08 2024-10-15 2024-10-31 2024-11-04 2024-11-05 2024-11-13 2024-11-14 
#>        192        384        960       2112       3744       4992       5073 
#> 2024-11-18 2024-11-20 2024-11-21 2024-12-16 2024-12-17 2025-01-14 2025-01-16 
#>       9201      11409      15246      18702      21402      21402      28871 
#> 2025-01-30 2025-01-31 2025-02-03 2025-02-05 2025-02-06 2025-02-07 2025-02-10 
#>      32135      34151      37511      40391      44615      47559      52248 
#> 2025-02-12 2025-02-13 2025-02-24 2025-02-27 2025-03-03 2025-03-17 2025-03-19 
#>      56568      60383      62303      66239      68733      70365      72861 
#> 2025-03-20 2025-03-24 2025-03-26 2025-03-27 2025-04-07 2025-04-09 2025-04-10 
#>      75549      78465      80899      83587      85123      88097      93467 
#> 2025-04-30 2025-05-01 2025-05-08 2025-05-14 2025-05-21 2025-05-28 2025-06-04 
#>      94715      98362     100666     105082     110362     111322     113146 
#> 2025-06-16 2025-06-25 2025-07-09 
#>     115286     117014     118438